Fixed a bunch more warnings, only have a few tricky ones left.

+++ decompile/General/LOAD/LOAD_A07_Callback_Podiums.c
@@ -3,6 +3,6 @@
 void DECOMP_LOAD_Callback_Podiums(struct LoadQueueSlot * lqs)
 {
   sdata->load_inProgress = 0;
-  data.podiumModel_podiumStands = lqs->ptrDestination;
+  data.podiumModel_podiumStands = (int)lqs->ptrDestination;
   return;
 }
